pyproject.toml
[tool.poetry]
name = "chatgptbot"
version = "0.1.0"
description = ""
authors = ["なるみ"]

[tool.poetry.dependencies]
python = "^3.8"
openai = "^0.27.1"
python-telegram-bot = "^20.1"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
yapf = "^0.32.0"
pytest = "^7.2.0"
codecov = "^2.1.12"
toml = "^0.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
